+++ app/src/main/java/com/luckyzyx/luckytool/hook/HookZygote.kt
@@ -2,10 +2,14 @@ package com.luckyzyx.luckytool.hook
 
 import com.highcapable.yukihookapi.hook.entity.YukiBaseHooker
 import com.luckyzyx.luckytool.hook.apps.android.DisableFlagSecureZygote
+import com.luckyzyx.luckytool.utils.tools.XposedPrefs
 
 class HookZygote : YukiBaseHooker() {
     override fun onHook() {
         //禁用FLAG_SECURE
-        loadHooker(DisableFlagSecureZygote())
+        if (prefs(XposedPrefs).getBoolean("disable_flag_secure",false)){
+            loadHooker(DisableFlagSecureZygote())
+        }
+
     }
 }