+++ app/src/main/java/com/luckyzyx/luckytool/hook/MainHook.kt
@@ -12,7 +12,6 @@ import com.luckyzyx.luckytool.hook.apps.CorePatch.CorePatchForR
 import com.luckyzyx.luckytool.hook.apps.CorePatch.CorePatchForS
 import com.luckyzyx.luckytool.hook.apps.CorePatch.CorePatchForSv2
 import com.luckyzyx.luckytool.hook.apps.CorePatch.CorePatchForT
-import com.luckyzyx.luckytool.hook.apps.Miscellaneous
 import com.luckyzyx.luckytool.hook.statusbar.*
 import com.luckyzyx.luckytool.utils.tools.XposedPrefs
 import de.robv.android.xposed.IXposedHookZygoteInit