+++ app/src/main/java/com/luckyzyx/luckytool/hook/apps/CorePatch/CorePatchForT.java
@@ -6,8 +6,7 @@ import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XposedHelpers;
 import de.robv.android.xposed.callbacks.XC_LoadPackage;
 
-@SuppressWarnings("RedundantThrows")
-public class CorePatchForT extends CorePatchForR {
+public class CorePatchForT extends CorePatchForSv2 {
     @Override
     public void handleLoadPackage(XC_LoadPackage.LoadPackageParam loadPackageParam) throws IllegalAccessException, InvocationTargetException, InstantiationException {
         super.handleLoadPackage(loadPackageParam);
@@ -20,11 +19,9 @@ public class CorePatchForT extends CorePatchForR {
                 new ReturnConstant(prefs, "downgrade", null));
 
         if (prefs.getBoolean("digestCreak", true) && prefs.getBoolean("UsePreSig", false)) {
-            Class<?> pmClass = findClass("com.android.server.pm.InstallPackageHelper", loadPackageParam.classLoader);
-            Class<?> pPClass = findClass("com.android.server.pm.parsing.pkg.ParsedPackage", loadPackageParam.classLoader);
-            XposedHelpers.findAndHookMethod(pmClass, "doesSignatureMatchForPermissions", String.class, pPClass, int.class, new XC_MethodHook() {
+            findAndHookMethod("com.android.server.pm.InstallPackageHelper", loadPackageParam.classLoader, "doesSignatureMatchForPermissions", String.class, "com.android.server.pm.parsing.pkg.ParsedPackage", int.class, new XC_MethodHook() {
                 @Override
-                protected void afterHookedMethod(MethodHookParam param) throws Throwable {
+                protected void afterHookedMethod(MethodHookParam param) {
                     //If we decide to crack this then at least make sure they are same apks, avoid another one that tries to impersonate.
                     if (param.getResult().equals(false)) {
                         String pPname = (String) XposedHelpers.callMethod(param.args[1], "getPackageName");