+++ app/src/main/java/com/luckyzyx/luckytool/ui/activity/ShortcutActivity.kt
@@ -4,7 +4,9 @@ import android.app.Activity
 import android.os.Bundle
 import com.joom.paranoid.Obfuscate
 import com.luckyzyx.luckytool.utils.tools.ShellUtils
+import com.luckyzyx.luckytool.utils.tools.jumpBatteryInfo
 import com.luckyzyx.luckytool.utils.tools.jumpRunningApp
+import kotlin.system.exitProcess
 
 @Obfuscate
 class ShortcutActivity : Activity(){
@@ -14,7 +16,9 @@ class ShortcutActivity : Activity(){
         when(shortcut?.getString("Shortcut")){
             "oplusGames" -> ShellUtils.execCommand("am start -n com.oplus.games/business.compact.activity.GameBoxCoverActivity", true)
             "processManager" -> jumpRunningApp(this)
+            "chargingTest" -> jumpBatteryInfo(this)
         }
         finishAffinity()
+        exitProcess(0)
     }
 }