+++ app/src/main/java/com/luckyzyx/luckytool/ui/fragment/OtherFragment.kt
@@ -1,18 +1,21 @@
 package com.luckyzyx.luckytool.ui.fragment
 
 import android.os.Bundle
+import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.view.inputmethod.EditorInfo
 import androidx.fragment.app.Fragment
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.google.android.material.switchmaterial.SwitchMaterial
+import com.google.android.material.textfield.TextInputEditText
+import com.google.android.material.textfield.TextInputLayout
+import com.google.android.material.textview.MaterialTextView
 import com.joom.paranoid.Obfuscate
 import com.luckyzyx.luckytool.R
 import com.luckyzyx.luckytool.databinding.FragmentOtherBinding
-import com.luckyzyx.luckytool.utils.tools.ShellUtils
-import com.luckyzyx.luckytool.utils.tools.jumpBatteryInfo
-import com.luckyzyx.luckytool.utils.tools.jumpEngineermode
-import com.luckyzyx.luckytool.utils.tools.jumpRunningApp
+import com.luckyzyx.luckytool.utils.tools.*
 
 @Obfuscate
 class OtherFragment : Fragment() {
@@ -21,7 +24,6 @@ class OtherFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         binding = FragmentOtherBinding.inflate(inflater)
-        setHasOptionsMenu(true)
         return binding.root
     }
 
@@ -42,16 +44,14 @@ class OtherFragment : Fragment() {
             MaterialAlertDialogBuilder(requireActivity())
                 .setCancelable(true)
                 .setItems(quicklist.toTypedArray()) { _, which ->
-                    when(which){
-                        0 -> { jumpEngineermode(requireActivity())
-                        }
-                        1 -> { jumpBatteryInfo(requireActivity())
-                        }
-                        2 -> { jumpRunningApp(requireActivity()) }
-                        3 -> { ShellUtils.execCommand("am start -n com.android.systemui/.DemoMode", true) }
-                        4 -> { ShellUtils.execCommand("am start -n com.oplus.logkit/.activity.MainActivity", true) }
-                        5 -> { ShellUtils.execCommand("am start -a com.android.settings.APPLICATION_DEVELOPMENT_SETTINGS", true) }
-                        6 -> { ShellUtils.execCommand("am start -n com.oplus.games/business.compact.activity.GameBoxCoverActivity", true) }
+                    when (which) {
+                        0 -> jumpEngineermode(requireActivity())
+                        1 -> jumpBatteryInfo(requireActivity())
+                        2 -> jumpRunningApp(requireActivity())
+                        3 -> ShellUtils.execCommand("am start -n com.android.systemui/.DemoMode", true)
+                        4 -> ShellUtils.execCommand("am start -n com.oplus.logkit/.activity.MainActivity", true)
+                        5 -> ShellUtils.execCommand("am start -a com.android.settings.APPLICATION_DEVELOPMENT_SETTINGS", true)
+                        6 -> ShellUtils.execCommand("am start -n com.oplus.games/business.compact.activity.GameBoxCoverActivity", true)
                     }
                 }
                 .show()
@@ -61,6 +61,71 @@ class OtherFragment : Fragment() {
         }.setOnCheckedChangeListener { buttonView, isChecked ->
             if (buttonView.isPressed) {
                 ShellUtils.execCommand("su -c service call SurfaceFlinger 1034 i32 ${if (isChecked) 1 else 0}", true)
+            }
+        }
+
+        binding.remoteAdbDebugTitle.text = getString(R.string.remote_adb_debug_title)
+        binding.remoteAdbDebugSummary.text = getString(R.string.remote_adb_debug_summary)
+        binding.remoteAdbDebug.setOnClickListener {
+            val getPort = ShellUtils.execCommand("getprop service.adb.tcp.port", true, true).successMsg
+            val getIP = ShellUtils.execCommand("ifconfig wlan0 | grep 'inet addr' | awk '{ print $2}' | awk -F: '{print $2}' 2>/dev/null", true, true).successMsg.let {
+                if (it == "") "IP" else it
+            }
+            val adbDialog = MaterialAlertDialogBuilder(requireActivity()).apply {
+                setCancelable(true)
+                setView(R.layout.layout_adb_dialog)
+            }.show()
+
+            adbDialog.findViewById<TextInputLayout>(R.id.adb_port_layout)?.apply {
+                hint = context.getString(R.string.adb_port)
+                isHintEnabled = true
+                isHintAnimationEnabled = true
+                isCounterEnabled = true
+                counterMaxLength = 6
+            }
+            val adbPort = adbDialog.findViewById<TextInputEditText>(R.id.adb_port)?.apply {
+                inputType = EditorInfo.TYPE_CLASS_NUMBER
+                setText(requireActivity().getString(OtherPrefs,"adb_port","6666"))
+            }
+            val adbTv = adbDialog.findViewById<MaterialTextView>(R.id.adb_tv)?.apply {
+                gravity = Gravity.CENTER
+                if (!(getPort == "" || getPort.toInt() == -1)){
+                    text = "adb connect $getIP:$getPort"
+                }
+            }
+
+            adbDialog.findViewById<SwitchMaterial>(R.id.adb_switch)?.apply {
+                text = context.getString(R.string.enable_remote_adb_debugging)
+                isChecked = !(getPort == "" || getPort.toInt() == -1)
+                setOnCheckedChangeListener { _, isChecked ->
+                    if (isChecked){
+                        val port = adbPort?.text.toString()
+                        if (port == "") {
+                            this.isChecked = false
+                            adbTv?.text = context.getString(R.string.adb_debug_port_cannot_null)
+                            return@setOnCheckedChangeListener
+                        }
+                        val commands = arrayOf(
+                            "setprop service.adb.tcp.port '$port'",
+                            "stop adbd",
+                            "killall -9 adbd 2>/dev/null",
+                            "start adbd"
+                        )
+                        ShellUtils.execCommand(commands,true)
+                        requireActivity().putString(OtherPrefs,"adb_port",port)
+                        adbTv?.text = "adb connect $getIP:$port"
+                    }else{
+                        val commands = arrayOf(
+                            "setprop service.adb.tcp.port -1",
+                            "stop adbd",
+                            "killall -9 adbd 2>/dev/null",
+                            "start adbd",
+                            "setprop service.adb.tcp.port ''"
+                        )
+                        ShellUtils.execCommand(commands,true)
+                        adbTv?.text = null
+                    }
+                }
             }
         }
     }